@@ -1,48 +1,31 @@
 {
-  "name": "auction-web",
-  "version": "0.1.0",
+  "name": "second-hand-auction-system-fe",
   "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.15.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "react": "^18.2.0",
-    "react-helmet": "^6.1.0",
-    "react-datepicker": "^4.5.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "6.0.2",
-    "react-router-hash-link": "^2.4.3",
-    "prop-types": "^15.8.1",
-    "react-rating-stars-component": "^2.2.0",
-    "react-select": "^5.2.1",
-    "react-alice-carousel": "^2.5.1",
-    "react-range": "^1.8.14",
-    "react-accessible-accordion": "^5.0.0",
-    "@tailwindcss/forms": "^0.5.7",
-    "dotenv": "^16.0.1"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "vite",
+    "dev": "vite",
     "build": "vite build",
-    "serve": "vite preview"
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": ["react-app", "react-app/jest"]
-  },
-  "browserslist": {
-    "production": [">0.2%", "not dead", "not op_mini all"],
-    "development": ["last 1 chrome version", "last 1 firefox version", "last 1 safari version"]
+  "dependencies": {
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.47",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.11"
   },
   "devDependencies": {
-    "prop-types": "^15.8.1",
-    "@vitejs/plugin-react": "^3.1.0",
-    "vite-tsconfig-paths": "3.6.0",
-    "vite": "^4.1.1",
-    "autoprefixer": "10.4.2",
-    "postcss": "8.4.8",
-    "tailwindcss": "3.4.6"
+    "@eslint/js": "^9.9.0",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "eslint": "^9.9.0",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+    "eslint-plugin-react-refresh": "^0.4.9",
+    "globals": "^15.9.0",
+    "vite": "^5.4.1"
   }
 }
-
-
-
