@@ -1,10 +1,6 @@
-module.exports = {
+export default {
   plugins: {
-    "tailwindcss/nesting": {},
     tailwindcss: {},
     autoprefixer: {},
   },
 };
-
-
-
